requirements.txt
PyFxA
requests
hkdf
